+++ pkg/nodes/lru/lru_test.go
@@ -14,9 +14,9 @@ func TestLRU_Add(t *testing.T) {
 	t.Parallel()
 
 	t.Run("adding with eviction", func(t *testing.T) {
-		fn := func(id0, id1 selectors.Field, rec0, rec1 float64) bool {
+		fn := func(id0, id1 selectors.Field, rec0, rec1 int64) bool {
 			evictted := 0
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				if expected, actual := id0, k; !expected.Equal(actual) {
 					t.Errorf("expected: %v, actual: %v", expected, actual)
 				}
@@ -54,8 +54,8 @@ func TestLRU_Add(t *testing.T) {
 	})
 
 	t.Run("adding sorts keys", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2, rec3 float64) bool {
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2, rec3 int64) bool {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				t.Fatal("failed if called")
 			}
 
@@ -88,8 +88,8 @@ func TestLRU_Get(t *testing.T) {
 	t.Parallel()
 
 	t.Run("get", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				t.Fatal("failed if called")
 			}
 
@@ -116,8 +116,8 @@ func TestLRU_Get(t *testing.T) {
 	})
 
 	t.Run("get sorts keys", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				t.Fatal("failed if called")
 			}
 
@@ -150,8 +150,8 @@ func TestLRU_Peek(t *testing.T) {
 	t.Parallel()
 
 	t.Run("peek", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				t.Fatal("failed if called")
 			}
 
@@ -178,8 +178,8 @@ func TestLRU_Peek(t *testing.T) {
 	})
 
 	t.Run("peek does not sorts keys", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				t.Fatal("failed if called")
 			}
 
@@ -212,8 +212,8 @@ func TestLRU_Contains(t *testing.T) {
 	t.Parallel()
 
 	t.Run("contains", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				t.Fatal("failed if called")
 			}
 
@@ -237,8 +237,8 @@ func TestLRU_Contains(t *testing.T) {
 	})
 
 	t.Run("does not contains", func(t *testing.T) {
-		fn := func(id0, id1, id2, id3 selectors.Field, rec0, rec1, rec2 float64) bool {
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		fn := func(id0, id1, id2, id3 selectors.Field, rec0, rec1, rec2 int64) bool {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				t.Fatal("failed if called")
 			}
 
@@ -266,9 +266,9 @@ func TestLRU_Remove(t *testing.T) {
 	t.Parallel()
 
 	t.Run("removes key value pair", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
 			evictted := 0
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				if expected, actual := id0, k; !expected.Equal(actual) {
 					t.Errorf("expected: %v, actual: %v", expected, actual)
 				}
@@ -308,7 +308,7 @@ func TestLRU_Pop(t *testing.T) {
 	t.Parallel()
 
 	t.Run("pop on empty", func(t *testing.T) {
-		onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 			t.Fatal("failed if called")
 		}
 
@@ -322,9 +322,9 @@ func TestLRU_Pop(t *testing.T) {
 	})
 
 	t.Run("pop", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
 			evictted := 0
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				if expected, actual := id0, k; !expected.Equal(actual) {
 					t.Errorf("expected: %v, actual: %v", expected, actual)
 				}
@@ -362,9 +362,9 @@ func TestLRU_Pop(t *testing.T) {
 	})
 
 	t.Run("pop results", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
 			evictted := 0
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				if expected, actual := id0, k; !expected.Equal(actual) {
 					t.Errorf("expected: %v, actual: %v", expected, actual)
 				}
@@ -404,9 +404,9 @@ func TestLRU_Purge(t *testing.T) {
 	t.Parallel()
 
 	t.Run("purge", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
 			evictted := 0
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				evictted += 1
 			}
 
@@ -446,8 +446,8 @@ func TestLRU_Keys(t *testing.T) {
 	t.Parallel()
 
 	t.Run("keys", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				t.Fatal("failed if called")
 			}
 
@@ -475,8 +475,8 @@ func TestLRU_Keys(t *testing.T) {
 	})
 
 	t.Run("keys after get", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				t.Fatal("failed if called")
 			}
 
@@ -510,9 +510,9 @@ func TestLRU_Dequeue(t *testing.T) {
 	t.Parallel()
 
 	t.Run("dequeue", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
 			evictted := 0
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				evictted += 1
 			}
 
@@ -531,7 +531,7 @@ func TestLRU_Dequeue(t *testing.T) {
 				t.Errorf("expected: %v, actual: %v", expected, actual)
 			}
 
-			got, err := l.Dequeue(func(key selectors.Field, value float64) error {
+			got, err := l.Dequeue(func(key selectors.Field, value int64) error {
 				return nil
 			})
 			if expected, actual := true, err == nil; expected != actual {
@@ -557,9 +557,9 @@ func TestLRU_Dequeue(t *testing.T) {
 	})
 
 	t.Run("dequeue with error", func(t *testing.T) {
-		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 float64) bool {
+		fn := func(id0, id1, id2 selectors.Field, rec0, rec1, rec2 int64) bool {
 			evictted := 0
-			onEviction := func(reason lru.EvictionReason, k selectors.Field, v float64) {
+			onEviction := func(reason lru.EvictionReason, k selectors.Field, v int64) {
 				evictted += 1
 			}
 
@@ -578,7 +578,7 @@ func TestLRU_Dequeue(t *testing.T) {
 				t.Errorf("expected: %v, actual: %v", expected, actual)
 			}
 
-			got, err := l.Dequeue(func(key selectors.Field, value float64) error {
+			got, err := l.Dequeue(func(key selectors.Field, value int64) error {
 				if key.Equal(id1) {
 					return errors.New("bad")
 				}
